screw this i hate this
turn timer complete, just need to figure out how to use date
